........S. [ZBXNEXT-8735] fixed warning

+++ src/libs/zbxhttp/http.c
@@ -624,7 +624,7 @@ next_attempt:
 		context->body.offset = 0;
 
 		if (0 != attempt_interval && 1 < context->max_attempts)
-			sleep(attempt_interval);
+			sleep((unsigned int)attempt_interval);
 	}
 	while (0 < --context->max_attempts);
